flask_twitter_clone/requirements.txt
# Core Flask Stack
Flask==3.1.0
Flask-Cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
python-dotenv==1.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40

# Templating and Web Server
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.1.8
itsdangerous==2.2.0

# Alembic (used by Flask-Migrate)
alembic==1.15.2
Mako==1.3.10

# FastAPI + async tools (optional but you had them)
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.2
uvloop==0.21.0
httptools==0.6.4
anyio==4.9.0
sniffio==1.3.1
httpcore==1.0.9
httpx==0.28.1
websockets==15.0.1

# Email + Validation
email_validator==2.2.0
aiosmtplib==4.0.1
dnspython==2.7.0

# OpenAI & Stripe
openai==1.76.0
stripe==12.1.0

# Data & Utilities
annotated-types==0.7.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
typing-inspection==0.4.0
typing_extensions==4.13.2
PyYAML==6.0.2
requests==2.32.3
urllib3==2.4.0
tqdm==4.67.1
distro==1.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
jiter==0.9.0
watchfiles==1.0.5
